requirements.txt
# pip install -r requirements.txt
requests
python-dotenv
# beautifulsoup4
# selenium
# httpx
# httpx[http2]